Move changePassword logic to respective component
